refactor : front modal 분리
